+++ apps/payments/nextjs/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @crossmint/client-sdk-nextjs-starter
 
+## 2.0.15
+
+### Patch Changes
+
+- @crossmint/client-sdk-react-ui@4.0.12
+
+## 2.0.14
+
+### Patch Changes
+
+- @crossmint/client-sdk-react-ui@4.0.11
+
+## 2.0.13
+
+### Patch Changes
+
+- @crossmint/client-sdk-react-ui@4.0.10
+
 ## 2.0.12
 
 ### Patch Changes