+++ packages/client/ui/react-ui/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @crossmint/client-sdk-react-ui
 
+## 4.0.12
+
+### Patch Changes
+
+- Updated dependencies [0535c3a]
+- Updated dependencies [612d768]
+  - @crossmint/client-sdk-auth@1.3.5
+  - @crossmint/wallets-sdk@1.0.8
+  - @crossmint/client-sdk-react-base@2.0.10
+
+## 4.0.11
+
+### Patch Changes
+
+- Updated dependencies [3a91c39]
+  - @crossmint/client-sdk-auth@1.3.4
+  - @crossmint/client-sdk-react-base@2.0.9
+
+## 4.0.10
+
+### Patch Changes
+
+- Updated dependencies [448d797]
+  - @crossmint/client-sdk-react-base@2.0.8
+
 ## 4.0.9
 
 ### Patch Changes